Clarified project types

+++ src/main/kotlin/Compiler.kt
@@ -169,7 +169,7 @@ class Compiler {
         )
         mutLines.replaceAll { if(it.matches("(;.+)".toRegex())) "\n" else it } //remove full line comments
         mutLines.replaceAll { if(it.contains(';')) it.substring(0 ..< it.indexOf(';')) else it } //remove eol comments
-        bytes += 127u; bytes += shortToTwoBytes((mutLines.size.toUShort() * 3u).toUShort())
+        bytes += 127u; bytes += shortToTwoBytes((mutLines.size * 3).toUShort())
         for(line in mutLines) {
             val split = line.split("[ ,]".toRegex())
             val (statement, arg) = split
